setup: removed yarn lock, only one package.json, setup SASS and react-router
