--- client/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "quiz",
-  "version": "0.1.0",
-  "private": true,
-  "devDependencies": {
-    "react-scripts": "1.1.5"
-  },
-  "dependencies": {
-    "react": "^16.5.0",
-    "react-dom": "^16.5.0"
-  },
-  "scripts": {
-    "start:react": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "proxy": "http://localhost:5000"
-}
