+++ src/App.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react';
+import logo from './logo.svg';
+import './App.css';
+import Routes from './routes';
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <header className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <h1 className="App-title">Welcome to React</h1>
+        </header>
+        <Routes />
+      </div>
+    );
+  }
+}
+
+export default App;
