+++ src/App.scss
@@ -8,8 +8,9 @@
 }
 
 .App-header {
-  background-color: #222;
+  background-color: rgb(34, 34, 34);
   height: 150px;
+  margin: 0;
   padding: 20px;
   color: white;
 }
