+++ src/TestComponent.js
@@ -1,8 +1,6 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
 
-class App extends Component {
+class Home extends Component {
   state = {
     response: ''
   };
@@ -24,19 +22,15 @@ class App extends Component {
 
     return body;
   };
+
   render() {
     return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          {this.state.response}
-        </p>
+      <div>
+        {this.state.response}
       </div>
-    );
+
+    )
   }
 }
 
-export default App;
+export default Home;
