+++ src/history.js
@@ -1,0 +1,4 @@
+// history.js
+import { createBrowserHistory } from 'history'
+
+export default createBrowserHistory();
